Implement rate limiting (#22)
Replace counting the number of requests per epoch with limiting RPS by
using the token bucket algorithm, as described
[here](https://github.com/subsquid/specs/blob/main/network-rfc/06_compute_units_allocation.md#rate-limiting).

On each query, if the limit has already been hit or the last token has
been used, the `retry_after_ms` hint is added to the response. It tells
the client that no more queries can be processed until that duration
elapses.